Ensure use-gold is honored for host tools
Chromium has a single use_gold GN setting for host and target. And we
set it by default with PACKAGECONFIG use-gold option.

But Chromium, asuming gold is used, just calls ld from the host tools
that could default to ld.bfd. To fix the issue, add the explicit
parameter -fuse-ld=gold to host linker calls.

This fixes an error passing --icf=none to the linker, an option that
does not exist in ld.bfd.